More entities

+++ src/main/java/entity/Event.java
@@ -1,6 +1,5 @@
 package entity;
 
-import java.math.BigInteger;
 
 /**
  * Created by Kevin Khanda on 7/23/2016.
@@ -8,20 +7,20 @@ import java.math.BigInteger;
  */
 public class Event implements DomainObject {
 
-    private BigInteger id;
+    private int id;
     private String name;
 
-    public Event(BigInteger id, String name) {
+    public Event(int id, String name) {
 
         this.id = id;
         this.name = name;
     }
 
-    public BigInteger getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(int id) {
         this.id = id;
     }
