+++ src/main/java/entity/Program.java
@@ -1,6 +1,5 @@
 package entity;
 
-import java.math.BigInteger;
 
 /**
  * Created by Kevin Khanda on 7/23/2016.
@@ -8,23 +7,23 @@ import java.math.BigInteger;
  */
 public class Program implements DomainObject {
 
-    private BigInteger id;
+    private int id;
     private String name;
     private String organisation;
     private String owner;
 
-    public Program(String name, BigInteger id, String organisation, String owner) {
+    public Program(String name, int id, String organisation, String owner) {
         this.name = name;
         this.id = id;
         this.organisation = organisation;
         this.owner = owner;
     }
 
-    public BigInteger getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(int id) {
         this.id = id;
     }
